DEX-148: Specify tick size in DEX -- order restrictions and matching rules specifications
