+++ dex/src/test/scala/com/wavesplatform/dex/settings/MatcherSettingsSpecification.scala
@@ -1,149 +1,18 @@
 package com.wavesplatform.dex.settings
 
-import cats.data.NonEmptyList
-import cats.implicits._
-import com.typesafe.config.{Config, ConfigFactory}
+import com.typesafe.config.Config
 import com.wavesplatform.common.utils.EitherExt2
 import com.wavesplatform.dex.api.OrderBookSnapshotHttpCache
 import com.wavesplatform.dex.queue.{KafkaMatcherQueue, LocalMatcherQueue}
 import com.wavesplatform.dex.settings.OrderFeeSettings.{DynamicSettings, FixedSettings, PercentSettings}
-import com.wavesplatform.settings.loadConfig
 import com.wavesplatform.state.diffs.produce
 import com.wavesplatform.transaction.assets.exchange.AssetPair
-import future.com.wavesplatform.transaction.assets.exchange.Implicits._
 import net.ceedubs.ficus.Ficus._
-import org.scalatest.{FlatSpec, Matchers}
+import org.scalatest.Matchers
 
 import scala.concurrent.duration._
-import scala.util.Try
 
-class MatcherSettingsSpecification extends FlatSpec with Matchers {
-
-  def getSettingByConfig(conf: Config): Either[String, MatcherSettings] =
-    Try(conf.as[MatcherSettings]("waves.dex")).toEither.leftMap(_.getMessage)
-
-  val correctOrderFeeStr: String =
-    s"""
-       |order-fee {
-       |  mode = percent
-       |  dynamic {
-       |    base-fee = 300000
-       |  }
-       |  fixed {
-       |    asset = WAVES
-       |    min-fee = 300000
-       |  }
-       |  percent {
-       |    asset-type = amount
-       |    min-fee = 0.1
-       |  }
-       |}
-       """.stripMargin
-
-  val correctDeviationsStr: String =
-    s"""
-       |max-price-deviations {
-       |  enable = yes
-       |  profit = 1000000
-       |  loss = 1000000
-       |  fee = 1000000
-       |}
-     """.stripMargin
-
-  val correctAllowedAssetPairsStr: String =
-    s"""
-       |allowed-asset-pairs = []
-     """.stripMargin
-
-  val correctOrderRestrictionsStr: String =
-    s"""
-       |order-restrictions = {}
-     """.stripMargin
-
-  val correctMatchingRulesStr: String =
-    s"""
-       |matching-rules = {}
-     """.stripMargin
-
-  def configWithSettings(orderFeeStr: String = correctOrderFeeStr,
-                         deviationsStr: String = correctDeviationsStr,
-                         allowedAssetPairsStr: String = correctAllowedAssetPairsStr,
-                         orderRestrictionsStr: String = correctOrderRestrictionsStr,
-                         matchingRulesStr: String = correctMatchingRulesStr): Config = {
-    val configStr =
-      s"""waves {
-      |  directory = /waves
-      |  dex {
-      |    account = 3Mqjki7bLtMEBRCYeQis39myp9B4cnooDEX
-      |    bind-address = 127.0.0.1
-      |    port = 6886
-      |    exchange-tx-base-fee = 300000
-      |    actor-response-timeout = 11s
-      |    snapshots-interval = 999
-      |    limit-events-during-recovery = 48879
-      |    make-snapshots-at-start = yes
-      |    snapshots-loading-timeout = 423s
-      |    start-events-processing-timeout = 543s
-      |    order-books-recovering-timeout = 111s
-      |    rest-order-limit = 100
-      |    price-assets = [
-      |      WAVES
-      |      8LQW8f7P5d5PZM7GtZEBgaqRPGSzS3DfPuiXrURJ4AJS
-      |      DHgwrRvVyqJsepd32YbBqUeDH4GJ1N984X8QoekjgH8J
-      |    ]
-      |    blacklisted-assets = ["a"]
-      |    blacklisted-names = ["b"]
-      |    blacklisted-addresses = [
-      |      3N5CBq8NYBMBU3UVS3rfMgaQEpjZrkWcBAD
-      |    ]
-      |    white-list-only = yes
-      |    allowed-order-versions = [11, 22]
-      |    order-book-snapshot-http-cache {
-      |      cache-timeout = 11m
-      |      depth-ranges = [1, 5, 333]
-      |    }
-      |    balance-watching-buffer-interval = 33s
-      |    events-queue {
-      |      type = "kafka"
-      |
-      |      local {
-      |        enable-storing = no
-      |        polling-interval = 1d
-      |        max-elements-per-poll = 99
-      |        clean-before-consume = no
-      |      }
-      |
-      |      kafka {
-      |        topic = "some-events"
-      |
-      |        consumer {
-      |          buffer-size = 100
-      |          min-backoff = 11s
-      |          max-backoff = 2d
-      |        }
-      |
-      |        producer {
-      |          enable = no
-      |          buffer-size = 200
-      |        }
-      |      }
-      |    }
-      |    process-consumed-timeout = 663s
-      |    $orderFeeStr
-      |    $deviationsStr
-      |    $allowedAssetPairsStr
-      |    $orderRestrictionsStr
-      |    $matchingRulesStr
-      |    exchange-transaction-broadcast {
-      |      broadcast-until-confirmed = yes
-      |      interval = 1 day
-      |      max-pending-time = 30 days
-      |    }
-      |  }
-      |}""".stripMargin
-
-    loadConfig(ConfigFactory.parseString(configStr))
-  }
+class MatcherSettingsSpecification extends BaseSettingsSpecification with Matchers {
 
   "MatcherSettings" should "read values" in {
 
@@ -399,156 +268,5 @@ class MatcherSettingsSpecification extends FlatSpec with Matchers {
         AssetPair.createAssetPair("WAVES", "ETH").get,
         AssetPair.createAssetPair("8LQW8f7P5d5PZM7GtZEBgaqRPGSzS3DfPuiXrURJ4AJS", "WAVES").get
       )
-  }
-
-  "Order restrictions" should "be validated" in {
-
-    def configStr(x: String): Config = configWithSettings(orderRestrictionsStr = x)
-
-    val nonEmptyCorrect =
-      """order-restrictions = {
-        | "WAVES-BTC": {
-        |   step-amount = 0.001
-        |   min-amount  = 0.001
-        |   max-amount  = 1000000
-        | },
-        | "ETH-USD": {
-        |   step-amount = 0.01
-        |   min-amount  = 0.05
-        |   max-amount  = 20000
-        | }
-        |}
-      """.stripMargin
-
-    val incorrectPairAndStepAmount =
-      """order-restrictions = {
-        | "WAVES-BTC": {
-        |   step-amount = -0.013
-        |   min-amount  = 0.001
-        |   max-amount  = 1000000
-        | },
-        | "ETH-;;;": {
-        |   step-amount = 0.01
-        |   min-amount  = 0.05
-        |   max-amount  = 20000
-        | }
-        |}
-      """.stripMargin
-
-    val incorrectMinAndMax =
-      """order-restrictions = {
-        | "WAVES-BTC": {
-        |   step-amount = 0.013
-        |   min-amount  = 0.001
-        |   max-amount  = 1000000
-        |   min-price   = 100
-        |   max-price   = 10
-        | },
-        | "ETH-WAVES": {
-        |   step-price = 0.14
-        |   max-price  = 17
-        | }
-        |}
-      """.stripMargin
-
-    withClue("default") {
-      getSettingByConfig(configStr("")).explicitGet().orderRestrictions shouldBe Map.empty
-    }
-
-    withClue("nonempty correct") {
-      getSettingByConfig(configStr(nonEmptyCorrect)).explicitGet().orderRestrictions shouldBe
-        Map(
-          AssetPair.createAssetPair("WAVES", "BTC").get ->
-            OrderRestrictionsSettings(
-              stepAmount = 0.001,
-              minAmount = 0.001,
-              maxAmount = 1000000,
-              stepPrice = OrderRestrictionsSettings.Default.stepPrice,
-              minPrice = OrderRestrictionsSettings.Default.minPrice,
-              maxPrice = OrderRestrictionsSettings.Default.maxPrice
-            ),
-          AssetPair.createAssetPair("ETH", "USD").get ->
-            OrderRestrictionsSettings(
-              stepAmount = 0.01,
-              minAmount = 0.05,
-              maxAmount = 20000,
-              stepPrice = OrderRestrictionsSettings.Default.stepPrice,
-              minPrice = OrderRestrictionsSettings.Default.minPrice,
-              maxPrice = OrderRestrictionsSettings.Default.maxPrice
-            )
-        )
-    }
-
-    withClue("incorrect pair and step amount") {
-      getSettingByConfig(configStr(incorrectPairAndStepAmount)) should produce(
-        "Invalid setting order-restrictions value: Can't parse asset pair 'ETH-;;;', " +
-          "Invalid setting order-restrictions.WAVES-BTC.step-amount value: -0.013 (required 0 < value)"
-      )
-    }
-
-    withClue("incorrect min and max") {
-      getSettingByConfig(configStr(incorrectMinAndMax)) should produce(
-        "Required order-restrictions.WAVES-BTC.min-price < order-restrictions.WAVES-BTC.max-price")
-    }
-  }
-
-  "Matching rules" should "be validated" in {
-    def configStr(x: String): Config = configWithSettings(matchingRulesStr = x)
-
-    val nonEmptyCorrect =
-      """matching-rules = {
-        |  "WAVES-8LQW8f7P5d5PZM7GtZEBgaqRPGSzS3DfPuiXrURJ4AJS": [
-        |    {
-        |      start-offset = 100
-        |      tick-size    = 0.002
-        |    },
-        |    {
-        |      start-offset = 500
-        |      tick-size    = 0.001
-        |    }
-        |  ]
-        |}
-      """.stripMargin
-
-    def incorrectRulesOrder(firstRuleOffset: Long, secondRuleOffset: Long): String =
-      s"""matching-rules = {
-        |  "WAVES-8LQW8f7P5d5PZM7GtZEBgaqRPGSzS3DfPuiXrURJ4AJS": [
-        |    {
-        |      start-offset = $firstRuleOffset
-        |      tick-size    = 0.002
-        |    },
-        |    {
-        |      start-offset = $secondRuleOffset
-        |      tick-size    = 0.001
-        |    }
-        |  ]
-        |}
-      """.stripMargin
-
-    withClue("default") {
-      getSettingByConfig(configStr("")).explicitGet().matchingRules shouldBe Map.empty
-    }
-
-    withClue("nonempty correct") {
-      getSettingByConfig(configStr(nonEmptyCorrect)).explicitGet().matchingRules shouldBe Map(
-        AssetPair.fromString("WAVES-8LQW8f7P5d5PZM7GtZEBgaqRPGSzS3DfPuiXrURJ4AJS").get ->
-          NonEmptyList[RawMatchingRules](
-            RawMatchingRules(100L, 0.002),
-            List(
-              RawMatchingRules(500L, 0.001)
-            )
-          )
-      )
-    }
-
-    withClue("incorrect rules order: 100, 100") {
-      getSettingByConfig(configStr(incorrectRulesOrder(100, 100))) should produce(
-        "Invalid setting matching-rules value: Rules should be ordered by offset, but they are: 100, 100")
-    }
-
-    withClue("incorrect rules order: 100, 88") {
-      getSettingByConfig(configStr(incorrectRulesOrder(100, 88))) should produce(
-        "Invalid setting matching-rules value: Rules should be ordered by offset, but they are: 100, 88")
-    }
   }
 }
